Moving python code into one place
